Split TLE-db functions to separate files, split transponder-db functions to separate files.
